Dataset Source File
Netflix is a popular streaming service that offers a vast catalog of movies, TV shows, and original contents. This dataset is a cleaned version of the original version which can be found here. The data consist of contents added to Netflix from 2008 to 2021.

Dataset consists of 10 columns and 8790 rows 

Source file Structure:

show_id - ID

type - Content(Tv show or Movie)

title - Title of Show

director - Director of the show or movie

country - Country of the show or movie

date_added - Date added to Netflix

release_year - Content release year

rating - Rating or restriction

duration - Duration of the content

listed_in - Genre of the content